Scaffold static export for public pages
- Move editor pages to /editor/* route (local dev only)
- Create public index and signs pages with getStaticProps
- Configure next.config.js for static export
- Exclude /editor and /api routes from static build

Dev mode: full editor at /editor/*
Static build: public pages only in /out
